requirements.txt
Django==2.2.1
djangorestframework==3.9.3
psycopg2==2.8.2
PyYAML==5.1
